Move DummyVerifier definition into dummy_types.hpp
Define "wrong" type in definition. This removes some duplicated code and
allows binding the type using pybind.
