+++ test/verification.cpp
@@ -17,19 +17,6 @@ using namespace arbitration_graphs_tests;
 
 using DummyPlaceboVerifier = verification::PlaceboVerifier<DummyCommand>;
 
-struct DummyVerifier {
-    // This could be made static here, but we want to challenge the compiler in deducing VerificationResultT.
-    // Unfortunately in such non-static cases VerificationResultT cannot be deduced
-    // and has to be passed on as template argument, see e.g. the DummyVerifierInPriorityArbitrator test
-    DummyResult analyze(const Time& /*time*/, const DummyCommand& data) const {
-        if (data == wrong_) {
-            return DummyResult{false};
-        }
-        return DummyResult{true};
-    };
-    std::string wrong_{"MidPriority"};
-};
-
 class CommandVerificationTest : public ::testing::Test {
 protected:
     DummyBehavior::Ptr testBehaviorHighPriority = std::make_shared<DummyBehavior>(false, false, "HighPriority");
@@ -94,7 +81,8 @@ TEST_F(CommandVerificationTest, PlaceboVerifier) {
 TEST_F(CommandVerificationTest, DummyVerifierInPriorityArbitrator) {
     using OptionFlags = PriorityArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult>::Option::Flags;
 
-    PriorityArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult> testPriorityArbitrator;
+    PriorityArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult> testPriorityArbitrator(
+        "PriorityArbitrator", DummyVerifier("MidPriority"));
 
     testPriorityArbitrator.addOption(testBehaviorHighPriority, OptionFlags::NO_FLAGS);
     testPriorityArbitrator.addOption(testBehaviorHighPriority, OptionFlags::NO_FLAGS);
@@ -159,7 +147,8 @@ TEST_F(CommandVerificationTest, DummyVerifierInPriorityArbitrator) {
 TEST_F(CommandVerificationTest, DummyVerifierInPriorityArbitratorWithFallback) {
     using OptionFlags = PriorityArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult>::Option::Flags;
 
-    PriorityArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult> testPriorityArbitrator;
+    PriorityArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult> testPriorityArbitrator(
+        "PriorityArbitrator", DummyVerifier("MidPriority"));
 
     testPriorityArbitrator.addOption(testBehaviorHighPriority, OptionFlags::NO_FLAGS);
     testPriorityArbitrator.addOption(testBehaviorHighPriority, OptionFlags::NO_FLAGS);
@@ -219,7 +208,8 @@ TEST_F(CommandVerificationTest, DummyVerifierInPriorityArbitratorWithFallback) {
 TEST_F(CommandVerificationTest, DummyVerifierInCostArbitrator) {
     using OptionFlags = CostArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult>::Option::Flags;
 
-    CostArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult> testCostArbitrator;
+    CostArbitrator<DummyCommand, DummyCommand, DummyVerifier, DummyResult> testCostArbitrator(
+        "CostArbitrator", DummyVerifier("MidPriority"));
 
     CostEstimatorFromCostMap::CostMap costMap{{"HighPriority", 0}, {"MidPriority", 0.5}, {"LowPriority", 1}};
     CostEstimatorFromCostMap::Ptr costEstimator = std::make_shared<CostEstimatorFromCostMap>(costMap);
